add-user.js
the event is called whenever a character is entered in the text field.

search for users by asynchronous communication each time a key is entered.

Receive the result of asynchronous communication from the argument of “done” function and create HTML to add to the view.

event fires when the add button is pressed.

adds the name of the user who clicked the add button to the chat member section and removes it from the search results.